refactored myserver -> app
