+++ backend/app/app.py
@@ -8,8 +8,8 @@ from fastapi import FastAPI
 from beanie import init_beanie
 from motor.motor_asyncio import AsyncIOMotorClient
 
-from myserver.config import CONFIG
-from myserver.models.user import User
+from app.config import CONFIG
+from app.models.user import User
 
 
 app = FastAPI()
@@ -20,4 +20,3 @@ async def app_init():
     """Initialize application services"""
     app.db = AsyncIOMotorClient(CONFIG.mongo_uri).account
     await init_beanie(app.db, document_models=[User])
-
