+++ backend/app/jwt.py
@@ -9,8 +9,8 @@ from fastapi.responses import JSONResponse
 from fastapi_jwt_auth import AuthJWT
 from fastapi_jwt_auth.exceptions import AuthJWTException
 
-from myserver.app import app
-from myserver.config import CONFIG
+from app.app import app
+from app.config import CONFIG
 
 
 @AuthJWT.load_config
